Fixed CA1034: Nested types should not be visible
