+++ ProcessCorrupt/SYSTEM_INFO.cs
@@ -1,0 +1,21 @@
+namespace RTCV.ProcessCorrupt
+{
+    using System;
+    using System.Runtime.InteropServices;
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct SYSTEM_INFO
+    {
+        public ushort processorArchitecture;
+        ushort reserved;
+        public uint pageSize;
+        public IntPtr minimumApplicationAddress;  // minimum address
+        public IntPtr maximumApplicationAddress;  // maximum address
+        public IntPtr activeProcessorMask;
+        public uint numberOfProcessors;
+        public uint processorType;
+        public uint allocationGranularity;
+        public ushort processorLevel;
+        public ushort processorRevision;
+    }
+}
